+++ packages/plugin-html-multi-response/CHANGELOG.md
@@ -1,8 +1,19 @@
 # @jspsych-contrib/plugin-html-multi-response
 
+## 1.0.2
+
+### Patch Changes
+
+- [#38](https://github.com/jspsych/jspsych-contrib/pull/38) [`f31ea01`](https://github.com/jspsych/jspsych-contrib/commit/f31ea012eb74eb4afe3e63af60d6e23bbbf3b9e8) Thanks [@richford](https://github.com/richford)! - First release of plugin-image-multi-response and plugin-image-swipe-response. Fix documentation link in plugin-html-multi-response
+
+## 1.0.1
+
+### Patch Changes
+
+- [#35](https://github.com/jspsych/jspsych-contrib/pull/35) [`6356a93`](https://github.com/jspsych/jspsych-contrib/commit/6356a932600af862d9fb4efb70a3f335ac1b7d96) Thanks [@richford](https://github.com/richford)! - Bug fix: Fixed an error in visual stimulation mode so that visual changes now depend on simulated data.
+
 ## 1.0.0
+
 ### Major Changes
-
-
 
 - [#33](https://github.com/jspsych/jspsych-contrib/pull/33) [`fb7ea33`](https://github.com/jspsych/jspsych-contrib/commit/fb7ea3303a046674f8261c7a84ebb80fdc4db9fb) Thanks [@richford](https://github.com/richford)! - Initial release